backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.18
httpx>=0.28.0
redis>=5.2.0
minio>=7.2.0
apscheduler>=3.10.0
structlog>=24.4.0
cryptography>=44.0.0
bibtexparser>=1.4.0
anthropic>=0.52.0
